Line graph full

+++ include/c_types/line_graph_rt.h
@@ -64,8 +64,8 @@ typedef struct {
     int64_t id;
     int64_t source;
     int64_t target;
-    float cost;
-    float reverse_cost;
+    double cost;
+    double reverse_cost;
 } Line_graph_rt;
 
 #endif  // INCLUDE_C_TYPES_LINE_GRAPH_RT_H_
